+++ docs/src/release_notes.md
@@ -9,6 +9,7 @@ These release notes adhere to the [keep a changelog](https://keepachangelog.com/
 *   Made calculation of p-values and standard errors optional and not executed by default in summarize methods [#65](https://github.com/dscolby/CausalELM.jl/issues/65)
 *   Removed redundant W argument for double machine learning, R-learning, and doubly robust estimation [#68](https://github.com/dscolby/CausalELM.jl/issues/68)
 ### Fixed
+*   Applying the weight trick for R-learning [#70](https://github.com/dscolby/CausalELM.jl/issues/70)
 
 ## Version [v0.6.0](https://github.com/dscolby/CausalELM.jl/releases/tag/v0.6.0) - 2024-06-15
 ### Added